requirements.txt
kivy
mysql-connector-python